Handle internal xrefs
There are many uses of the @see tags in the doc markup that assume you can link directly to another method/property without qualification.  So the doc generator now tries to take this into account.

- If a @see tag begins with a period or hashmark, assume it's a link to another method/property on that page
- If the @see tag link contains the name of the page within itself, also treat it as an internal link.  (So if Crafty.viewport.pan links to  Crafty.viewport.zoom, it is correctly interpreted as linking within that page.)

Some of this stuff should maybe be handled more directly by the script that parses the source, but this is a relatively small patch that immediately fixes several broken links.